refactor(dash): use Map() in menu array
