+++ dashboard/src/router/routes.ts
@@ -3,7 +3,7 @@ import { RouteRecordRaw } from 'vue-router';
 const routes: RouteRecordRaw[] = [
     {
         path: '/',
-        component: () => import('src/layouts/MainLayout.vue'),
+        component: () => import('src/layouts/MainLayout/Index.vue'),
         children: [
             {
                 path: '',
